+++ crates/conclave-server/src/main.rs
@@ -23,14 +23,13 @@ struct Cli {
 async fn main() -> anyhow::Result<()> {
     tracing_subscriber::fmt()
         .with_env_filter(
-            tracing_subscriber::EnvFilter::try_from_default_env()
-                .unwrap_or_else(|_| {
-                    if cfg!(debug_assertions) {
-                        "conclave_server=info".into()
-                    } else {
-                        "conclave_server=warn".into()
-                    }
-                }),
+            tracing_subscriber::EnvFilter::try_from_default_env().unwrap_or_else(|_| {
+                if cfg!(debug_assertions) {
+                    "conclave_server=info".into()
+                } else {
+                    "conclave_server=warn".into()
+                }
+            }),
         )
         .init();